Create new tmp files for upload gallery images on Android
Android overwrites it's own temporary image when you request more
than one image that's resized from the original. This means that
we upload the same image multiple times (even though it looks ok
in the webview). This adds a new function renameFile that copies
the temp file Android creates to a new one with a unique name, so
that it doesn't get overwritten when we request another one. The
function is then called when we get an image from the gallery
(on the Android platform only) to save all of the images.