Upgrade to Cordova 6.1.1 (Android 5.1.1/IOS 4.1.0)
This upgrades the base Cordova CLI to v6.1.1, the Android platform to v5.1.1
and the IOS platform to v4.1.0, as well as all of the plugins to their latest
versions.

Changes required/enabled by this are:
- Platforms/Plugins and their versions are now specified in config.xml
- <access> tags in config.xml needed a protocol wildcard to make the whitelist
  work with IOS9's App Transport Security (because many things are not https,
  IOS would block access to them by default).
- <preference name="AndroidPersistentFileLocation" value="Compatibility" />
  added to config.xml for the file plugin to be able to access existing files

I haven't added a Content Security Policy yet (https://github.com/mysociety/zurich_mobile/issues/14)
as there are other issues we could fix first to make that easier.