Fix iOS crash caused by missing NSPhotoLibraryAddUsageDescription
Without a string describing why photos were being saved to the iOS
photo library in the app's Info.plist, the app would crash as soon as
a photo was taken. Not quite sure how this wasn't picked up by local
testing or the automated and manual app store review process.

This change requires a newer version of the camera plugin, so I took the
opportunity to upgrade all Cordova plugins at the same time.